innotech-platform/backend/requirements-lambda.txt
fastapi==0.115.14
uvicorn==0.35.0
sqlalchemy==2.0.41
psycopg2-binary
alembic==1.16.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
boto3==1.39.3
python-dotenv==1.1.1
email-validator==2.2.0
mangum==0.18.0